requirements/prod.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==4.2.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0.1
sqlparse==0.4.2
tzdata==2022.2
Unipath==1.1
uritemplate==4.1.1
